更新

+++ scripts/downloadPg.sh
@@ -18,12 +18,18 @@ if [ "$LOCAL" = "${REMOTE}" ]; then
   exit 2
 fi
 
-echo "download ${REMOTE}" && \
-wget https://gitlab.com/power0721/pg/-/raw/main/pg.zip -O /data/pg.zip && \
-echo "unzip file" && \
-rm -rf /www/pg/* && unzip -q -o /data/pg.zip -d /www/pg && \
-echo "save version" && \
-echo -n ${REMOTE} > /data/pg_version.txt && \
-echo "sync files" && \
-[ -d /data/pg ] && \
-cp -r /data/pg/* /www/pg/
+LOCAL_NUM=$(echo $LOCAL | tr -d '-')
+REMOTE_NUM=$(echo $REMOTE | tr -d '-')
+if (( REMOTE_NUM > LOCAL_NUM )); then
+  echo "download ${REMOTE}" && \
+  wget https://gitlab.com/power0721/pg/-/raw/main/pg.zip -O /data/pg.zip && \
+  echo "unzip file" && \
+  rm -rf /www/pg/* && unzip -q -o /data/pg.zip -d /www/pg && \
+  echo "save version" && \
+  echo -n ${REMOTE} > /data/pg_version.txt && \
+  echo "sync files" && \
+  [ -d /data/pg ] && \
+  cp -r /data/pg/* /www/pg/
+fi
+
+
